requirements.txt
fastapi==0.85.1
python-dotenv==1.0.1
uvicorn==0.18.3
SQLAlchemy==1.4.42
psycopg2==2.9.4
fastapi-mail
requests==2.28.1